Rename filenames in models/ to match classnames
