+++ keras_nlp/models/bert/bert_classifier_test.py
@@ -20,9 +20,9 @@ from absl.testing import parameterized
 from tensorflow import keras
 
 from keras_nlp.models.bert.bert_backbone import BertBackbone
-from keras_nlp.models.bert.bert_preprocessing import BertPreprocessor
-from keras_nlp.models.bert.bert_preprocessing import BertTokenizer
-from keras_nlp.models.bert.bert_tasks import BertClassifier
+from keras_nlp.models.bert.bert_classifier import BertClassifier
+from keras_nlp.models.bert.bert_preprocessor import BertPreprocessor
+from keras_nlp.models.bert.bert_preprocessor import BertTokenizer
 
 
 class BertClassifierTest(tf.test.TestCase, parameterized.TestCase):
