+++ keras_nlp/models/distil_bert/distil_bert_classifier_test.py
@@ -19,8 +19,10 @@ import tensorflow as tf
 from absl.testing import parameterized
 from tensorflow import keras
 
-from keras_nlp.models.distilbert.distil_bert_models import DistilBert
-from keras_nlp.models.distilbert.distil_bert_tasks import DistilBertClassifier
+from keras_nlp.models.distil_bert.distil_bert_backbone import DistilBert
+from keras_nlp.models.distil_bert.distil_bert_classifier import (
+    DistilBertClassifier,
+)
 
 
 class DistilBertClassifierTest(tf.test.TestCase, parameterized.TestCase):
