+++ keras_nlp/models/roberta/roberta_preprocessor_test.py
@@ -20,8 +20,8 @@ import tensorflow as tf
 from absl.testing import parameterized
 from tensorflow import keras
 
-from keras_nlp.models.roberta.roberta_preprocessing import RobertaPreprocessor
-from keras_nlp.models.roberta.roberta_preprocessing import RobertaTokenizer
+from keras_nlp.models.roberta.roberta_preprocessor import RobertaPreprocessor
+from keras_nlp.models.roberta.roberta_preprocessor import RobertaTokenizer
 
 
 class RobertaTokenizerTest(tf.test.TestCase, parameterized.TestCase):
