+++ keras_nlp/models/xlm_roberta/xlm_roberta_backbone.py
@@ -19,14 +19,14 @@ import os
 
 from tensorflow import keras
 
-from keras_nlp.models.roberta import roberta_models
+from keras_nlp.models.roberta import roberta_backbone
 from keras_nlp.models.xlm_roberta.xlm_roberta_presets import backbone_presets
 from keras_nlp.utils.python_utils import classproperty
 from keras_nlp.utils.python_utils import format_docstring
 
 
 @keras.utils.register_keras_serializable(package="keras_nlp")
-class XLMRoberta(roberta_models.Roberta):
+class XLMRoberta(roberta_backbone.Roberta):
     """XLM-RoBERTa encoder.
 
     This network implements a bi-directional Transformer-based encoder as
