+++ keras_nlp/models/xlm_roberta/xlm_roberta_preprocessor_test.py
@@ -22,10 +22,10 @@ import tensorflow as tf
 from absl.testing import parameterized
 from tensorflow import keras
 
-from keras_nlp.models.xlm_roberta.xlm_roberta_preprocessing import (
+from keras_nlp.models.xlm_roberta.xlm_roberta_preprocessor import (
     XLMRobertaPreprocessor,
 )
-from keras_nlp.models.xlm_roberta.xlm_roberta_preprocessing import (
+from keras_nlp.models.xlm_roberta.xlm_roberta_preprocessor import (
     XLMRobertaTokenizer,
 )
